- rtkpos class for RTK based on pppos
